protecting the routes

+++ routes/web.php
@@ -29,25 +29,25 @@ use Illuminate\Support\Facades\Route;
 // Show all listings
 Route::get('/', [ListingController::class, 'index']);
 // Show a create form
-Route::get('/listings/create', [ListingController::class, 'create']);
+Route::get('/listings/create', [ListingController::class, 'create'])->middleware('auth');
 // Store a new listing
-Route::post('/listings', [ListingController::class, 'store']);
+Route::post('/listings', [ListingController::class, 'store'])->middleware('auth');
 // Show edit listing form
-Route::get('/listings/{listing}/edit', [ListingController::class, 'edit']);
+Route::get('/listings/{listing}/edit', [ListingController::class, 'edit'])->middleware('auth');
 // Edit/Update a listing
-Route::put('listings/{listing}', [ListingController::class, 'update']);
+Route::put('listings/{listing}', [ListingController::class, 'update'])->middleware('auth');
 // Delete a listing
-Route::delete('listings/{listing}', [ListingController::class, 'destroy']);
+Route::delete('listings/{listing}', [ListingController::class, 'destroy'])->middleware('auth');
 // Show a single listing
 Route::get('/listings/{listing}', [ListingController::class, 'show']);
 
 // Show register from
-Route::get('/register', [UserController::class, 'create']);
+Route::get('/register', [UserController::class, 'create'])->middleware('guest');
 // Store/Create a new user
 Route::post('/users', [UserController::class, 'store']);
 // Logs user our
-Route::post('/logout', [UserController::class, 'logout']);
+Route::post('/logout', [UserController::class, 'logout'])->middleware('auth');
 // Show login form
-Route::get('/login', [UserController::class, 'login']);
+Route::get('/login', [UserController::class, 'login'])->name('login')->middleware('guest');;
 // Login user
 Route::post('/users/authenticate', [UserController::class, 'authenticate']);
